Added warnings to compile script and fix some

+++ include/testing/isl-point-gen-helpers.h
@@ -2,19 +2,19 @@
 
 #include "isl-noexceptions.h"
 
-static int point_count = 0;
+static size_t point_count = 0;
 
 isl::point
 get_point_with_coordinates(isl::space point_space)
 {
     isl::point point_out(point_space);
     isl::ctx ctx = point_space.get_ctx();
-    for (int i = 0; i < point_space.dim(isl::dim::set); ++i)
+    for (size_t i = 0; i < point_space.dim(isl::dim::set); ++i)
     {
         isl::val coord_val(ctx, i + point_count);
         point_out = point_out.set_coordinate_val(isl::dim::set, i, coord_val);
     }
-    for (int i = 0; i < point_space.dim(isl::dim::param); ++i)
+    for (size_t i = 0; i < point_space.dim(isl::dim::param); ++i)
     {
         isl::val coord_val(ctx, i + point_count);
         point_out = point_out.set_coordinate_val(isl::dim::param, i, coord_val);
